HUGE REFACTORING. MAY CAUSE BACKWARDS INCOMPATIBILITY - refactored sizes of grid points, use smallest size possible, and unsigned instead of signed, therefore new recordings are not backwards compatible and old ones can't be read - remove invisible rows and replace the calculation, that needed them with others (game over on top) - refactored single_player game to be easily expandable, so that it only implements scene stuff and not more, so that game can be used in multiplayer games more easily -  create input factory to create correct input - move inputs into separate file per platform and type - remove  active_scenes from service_provider as it's not needed anymore ( game over scene return s StopUpdating) - add StopUpdating to all fullscreen-scenes, so that it is the last updated -  remove input folder and move them to platform or manager - AbstractRect and AbstractPoint support now smaller types than int, since those need explicit casting after any operation (+, - etc.) - change settings, this is still WIP -  changed clocksource to use local definition of elapsed_time - remove std::exception without message in places where not using spdlog::error

+++ src/game/tetrion_snapshot.hpp
@@ -6,6 +6,7 @@
 #include <span>
 #include <vector>
 
+//TODO: this needs versioning, since now we changed some sizes of some types!!
 struct TetrionSnapshot final {
 public:
     using Level = decltype(Tetrion::m_level);
@@ -22,7 +23,7 @@ private:
 
 public:
     using MinoCount = u64;
-    using Coordinate = i32;
+    using Coordinate = u8;
     using MinoType = u8;
 
     TetrionSnapshot(
